feat: refactor external ports to use 'api' package structure
