+++ user/src/main/java/com/lxp/user/application/port/api/dto/UserAuthResponse.java
@@ -1,4 +1,4 @@
-package com.lxp.user.application.port.provided.dto;
+package com.lxp.user.application.port.api.dto;
 
 import com.lxp.user.domain.user.model.entity.User;
