+++ user/src/main/java/com/lxp/user/application/port/api/external/ExternalUserInfoPort.java
@@ -1,6 +1,6 @@
-package com.lxp.user.application.port.provided.external;
+package com.lxp.user.application.port.api.external;
 
-import com.lxp.user.application.port.provided.dto.UserInfoResponse;
+import com.lxp.user.application.port.api.dto.UserInfoResponse;
 
 import java.util.Optional;
